pyproject.toml
# =============================================================================
# RNV Text Transformer — Project Metadata & Build Configuration
# =============================================================================
# Version is read dynamically from utils/config.py so that APP_VERSION remains
# the single source of truth for the application version.

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# =============================================================================
# Project metadata (PEP 621)
# =============================================================================

[project]
name = "rnv-text-transformer"
dynamic = ["version"]
description = "A professional-grade desktop text transformation suite built with PyQt6."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Chris Vizirov", email = "contact@rnvsoftware.dev" },
]
keywords = [
    "text-processing",
    "pyqt6",
    "desktop-app",
    "batch-processing",
    "regex",
    "file-conversion",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: X11 Applications :: Qt",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing",
    "Topic :: Text Processing :: Filters",
    "Topic :: Utilities",
    "Typing :: Typed",
]

# ---- Runtime dependencies ---------------------------------------------------
# Kept in sync with requirements.txt
dependencies = [
    "PyQt6>=6.6.0",
    "Pillow>=10.0.0",
    "python-docx>=1.1.0",
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
    "striprtf>=0.0.26",
    "chardet>=5.0.0",
    "watchdog>=3.0.0",
]

# ---- Optional dependency groups --------------------------------------------
[project.optional-dependencies]
dev = [
    "mypy>=1.8",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/RNVSoftware/rnv-text-transformer"
Repository = "https://github.com/RNVSoftware/rnv-text-transformer"
Issues = "https://github.com/RNVSoftware/rnv-text-transformer/issues"

# ---- Entry points -----------------------------------------------------------
# Console script for the CLI
[project.scripts]
rnv-transform = "cli.rnv_transform:main"

# GUI script (no console window on Windows)
[project.gui-scripts]
rnv-text-transformer = "RNV_Text_Transformer:main"


# =============================================================================
# Hatch build configuration
# =============================================================================

# Dynamic version extraction from utils/config.py — keeps APP_VERSION as the
# single source of truth. The regex matches: APP_VERSION: str = "X.Y.Z"
[tool.hatch.version]
path = "utils/config.py"
pattern = '^APP_VERSION:\s*str\s*=\s*"(?P<version>[^"]+)"'

[tool.hatch.build.targets.wheel]
packages = ["core", "ui", "utils", "cli"]

[tool.hatch.build.targets.wheel.force-include]
"RNV_Text_Transformer.py" = "RNV_Text_Transformer.py"

[tool.hatch.build.targets.sdist]
include = [
    "RNV_Text_Transformer.py",
    "core/",
    "ui/",
    "utils/",
    "cli/",
    "resources/",
    "README.md",
    "LICENSE",
    "requirements.txt",
]


# =============================================================================
# Tool configuration
# =============================================================================

[tool.mypy]
python_version = "3.10"
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
check_untyped_defs = true
# PyQt6 stubs can be noisy; ignore when helpful
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
